chore: renamed route blueprint into api/projects/:pid/issues, and adjusted methods accordingly
