+++ app/models/__init__.py
@@ -1,12 +1,12 @@
 from .projects import fetch_projects, add_project, delete_project_by_id, update_project_name
-from .project_errors import fetch_data_by_project, delete_data_by_project, fetch_error
+from .project_issues import fetch_issues_by_project, delete_issues_by_project, fetch_error
 
 __all__ = [
     "fetch_projects",
     "add_project",
     "delete_project_by_id",
     "update_project_name",
-    "fetch_data_by_project",
-    "delete_data_by_project",
+    "fetch_issues_by_project",
+    "delete_issues_by_project",
     "fetch_error",
 ]
