+++ app/routes/__init__.py
@@ -1,2 +1,2 @@
 from app.routes.projects import bp as projects_bp
-from app.routes.project_errors import bp as errors_bp
+from app.routes.project_issues import bp as issues_bp
