+++ app/routes/project_issues.py
@@ -1,12 +1,12 @@
 from flask import jsonify, request
 from flask import Blueprint
 from app.models import (
-    fetch_data_by_project,
-    delete_data_by_project,
+    fetch_issues_by_project,
+    delete_issues_by_project,
     fetch_error,
 )
 
-bp = Blueprint('project_errors', __name__)
+bp = Blueprint('project_issues', __name__)
 
 @bp.route('/', methods=['GET'])
 def get_errors(pid):
@@ -17,7 +17,7 @@ def get_errors(pid):
     resolved = request.args.get('resolved', None)
 
     try:
-        data = fetch_data_by_project(pid, page, limit, handled, time, resolved)
+        data = fetch_issues_by_project(pid, page, limit, handled, time, resolved)
         return jsonify(data), 200
     except Exception as e:
         return jsonify({"message": "Failed to fetch data", "error": str(e)}), 500
@@ -25,7 +25,7 @@ def get_errors(pid):
 @bp.route('/', methods=['DELETE'])
 def delete_errors(pid):
     try:
-        success = delete_data_by_project(pid)
+        success = delete_issues_by_project(pid)
         if success:
             return '', 204
         else: 
@@ -33,7 +33,7 @@ def delete_errors(pid):
     except Exception as e:
         return jsonify({"message": "Failed to delete errors", "error": str(e)}), 500
     
-@bp.route('/<eid>', methods=['GET'])
+@bp.route('/errors/<eid>', methods=['GET'])
 def get_error(_, eid):
     try:
         error = fetch_error(eid)
@@ -43,3 +43,4 @@ def get_error(_, eid):
             return jsonify({"message": "Error not found"}), 404
     except Exception as e:
         return jsonify({"message": "Failed to retrieve error", "error": str(e)}), 500
+
